// src/generated/models/Cb_todosubtasksModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cb_todosubtasksstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cb_todosubtasksstatecode = keyof typeof Cb_todosubtasksstatecode;
export const Cb_todosubtasksstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cb_todosubtasksstatuscode = keyof typeof Cb_todosubtasksstatuscode;

export interface Cb_todosubtasksBase {
  cb_isdone?: boolean;
  cb_sortorder?: number;
  "cb_todotask_cb_todosubtask_task@odata.bind": string;
  cb_title: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cb_todosubtasksstatecode;
  statuscode?: Cb_todosubtasksstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cb_todosubtasks extends Cb_todosubtasksBase {
  cb_isdonename?: string;
  cb_todosubtaskid: string;
  createdon?: string;
  modifiedon?: string;
  ownerid: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cb_task?: object;
  _cb_task_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}